Move DefaultEmpty
