+++ wire-library/wire-runtime-swift/src/main/swift/propertyWrappers/StringEncoded.swift
@@ -88,12 +88,6 @@ extension StringEncoded : Encodable {
     }
 }
 
-extension StringEncoded : EmptyInitializable where Value: EmptyInitializable {
-    public init() {
-        self.init(wrappedValue: Value())
-    }
-}
-
 extension StringEncoded : Equatable where Value : Equatable {
 }
